refactor(CodeEditor): replace createHighlightExtension with useHighlightExtension for syntax highlighting
