+++ packages/code-editor/src/CodeEditor/codeMirrorExtensions/index.ts
@@ -1,2 +1,1 @@
-export { createHighlightExtension } from './createHighlightExtension';
 export { createThemeExtension } from './createThemeExtension';
